Garbage collection for PBFT & make code more DRY
As soon as we commit locally, we can remove the entry from our log. This
significantly minimizes the memory usage as otherwise the log constantly
grows.

Additionally, this commit improves a bit the code structure of PBFT.
Main change here is the repetetive workflow to check for predicates that is now
outsourced to a separate function.